fix: useRepositoryIssues hook and design
