+++ cra-rxjs-styled-components/src/components/pagination/Pagination.tsx
@@ -1,0 +1,10 @@
+import { PaginationContainer } from './Pagination.styles';
+
+export default function Pagination() {
+  return (
+    <PaginationContainer>
+      <span className="prev">Previous</span>
+      <span className="next">Next</span>
+    </PaginationContainer>
+  );
+}
