+++ cra-rxjs-styled-components/src/components/pagination/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './Pagination';
