+++ cra-rxjs-styled-components/src/components/repo-issues/IssueTabHeader/IssueTabHeader.tsx
@@ -1,7 +1,7 @@
 import { useCallback, useState } from 'react';
 import DetailsDropdown from '../../DetailsDropdown';
 import CorrectIcon from '../../icons/CorrectIcon';
-import PullRequestIcon from '../../icons/PullRequestIcon';
+import OpenIssueIcon from '../../icons/OpenIssueIcon';
 import type { IssueTabValues } from '../../../types/types';
 
 import { Container, StatusLabel, StatusTab } from './IssueTabHeader.style';
@@ -37,7 +37,7 @@ export default function IssueTabHeader(props: Props) {
           onClick={() => changeTab('open')}
           active={activeTab === 'open'}
         >
-          <PullRequestIcon />
+          <OpenIssueIcon />
           <span>{openCount}</span>
           <span>Open</span>
         </StatusLabel>
